+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - get a self service token for a given contract
 - get a specific customer by id
+- get and delete rated items
 
 ### Changed
 - update version of jruby to 9.1.5.0